# openapi.yaml
components:
  schemas:
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: A service that logs time intervals
  title: Timelog
  version: 0.1.0
openapi: 3.1.0
paths:
  /:
    get:
      description: Return 200 just to show that the service is running.
      operationId: health__get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: health check
  /projects:
    post:
      description: 'Add a project.


        Returns 409 if the project already exists.'
      operationId: add_project_projects_post
      parameters:
      - in: query
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '201':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Add a new project
  /projects/{name}/activate:
    patch:
      description: 'Activates a project.


        All other active projects will be deactivated.


        Returns 404 if the project doesn''t exist.'
      operationId: activate_project_projects__name__activate_patch
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Activates a project
  /projects/{name}/deactivate:
    patch:
      description: 'Deactivates a project.


        Returns 404 if the project doesn''t exist.'
      operationId: deactivate_project_projects__name__deactivate_patch
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Deactivates a project
  /projects/{name}/duration:
    get:
      description: 'Retrieves project duration in hours, rounded to one decimal place.


        Returns 404 if the project doesn''t exist.'
      operationId: get_project_duration_projects__name__duration_get
      parameters:
      - in: path
        name: name
        required: true
        schema:
          title: Name
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                title: Response Get Project Duration Projects  Name  Duration Get
                type: number
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Retrieves project duration in hours, rounded to one decimal place
  /rpc/deactivate_all:
    patch:
      description: Deactivates all active projects.
      operationId: deactivate_all_rpc_deactivate_all_patch
      responses:
        '204':
          description: Successful Response
      summary: Deactivates all active projects